+++ CHANGELOG.md
@@ -1,10 +1,19 @@
+## [Version 0.3.1] - 2017-08-27
+### Added
+- Reparses open documents when configuration changes
+
+### Fixed
+- The parser now substitutes variables in `amxxpawn.compiler.includePaths` too
+
+
 ## [Version 0.3.0] - 2017-08-27
-## Added
+### Added
 - Substitution variables are now allowed in settings containing paths
 - More diagnostics to the **#include** statement parser
 
-## Fixed
+### Fixed
 - **#include** statements parser now provides links and diagnostics with a correct character range
+
 
 ## [Version 0.2.1] - 2017-08-27
 ### Added
@@ -16,6 +25,7 @@
 - Properly handle multiple multiline comments on the same line
 - Properly handle multiple braces on the same line
 
+
 ## [Version 0.2.0] - 2017-08-25
 ### Added
 - Document symbol lookup (`Ctrl+Shift+O`) - Easy way to search and navigate to any symbol in the currently opened document
@@ -24,22 +34,27 @@
 ### Fixed
 - Included dependencies are now properly managed, no more data leaks
 
+
 ## [Version 0.1.0] - 2017-08-25
 ### Added
 - 'Compile Plugin Local' command which searches for amxxpc executable in input file's path
+
 
 ## [Version 0.0.5] - 2017-08-25
 ### Fixed
 - Fixed amxxpawn.compiler.outputType === 'path'
 - AMXXPC Output panel now gets focus on compilation
 
+
 ## [Version 0.0.4] - 2017-08-25
 ### Fixed
 - Whitespaces between function's tag and identifier would break the parser
 
+
 ## [Version 0.0.3] - 2017-08-25
 ### Fixed
 - Crash when parsing functions with no storage specifiers
+
 
 ## [Version 0.0.1] - 2017-08-25
 - Initial release